+++ example/lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:example/constants.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
+import 'package:multiple_search_selection/helpers/create_options.dart';
 import 'package:multiple_search_selection/multiple_search_selection.dart';
 
 void main() {
@@ -45,7 +46,7 @@ class _MyHomePageState extends State<MyHomePage> {
       appBar: AppBar(
         title: Text(widget.title),
       ),
-      body: MultipleSearchSelection<Country>(
+      body: MultipleSearchSelection<Country>.creatable(
         title: Padding(
           padding: const EdgeInsets.all(12.0),
           child: Text(
@@ -55,6 +56,21 @@ class _MyHomePageState extends State<MyHomePage> {
               fontWeight: FontWeight.bold,
             ),
           ),
+        ),
+        onItemAdded: (c) {},
+
+        createOptions: CreateOptions(
+          createItem: (text) {
+            return Country(name: text, iso: text);
+          },
+          createItemBuilder: (text) => Align(
+            alignment: Alignment.centerLeft,
+            child: Padding(
+              padding: const EdgeInsets.all(8.0),
+              child: Text('Create "$text"'),
+            ),
+          ),
+          pickCreatedItem: true,
         ),
         items: countries, // List<Country>
         fieldToCheck: (c) {